Split classes.py into separate modules. Create gui.py, session.py, tree.py, gedcom.py. Update version to 1.0.2-rc.2. Add classes directory to setup.cfg.
