+++ getmyancestors/__init__.py
@@ -3,4 +3,4 @@
 from . import getmyancestors
 from . import mergemyancestors
 
-__version__ = "1.0.2-rc.1"
+__version__ = "1.0.2-rc.2"
