+++ getmyancestors/classes/constants.py
