+++ getmyancestors/fstogedcom.py
@@ -2,44 +2,17 @@
 # coding: utf-8
 
 # global imports
-import re
 import os
 import sys
-import time
-import asyncio
-import tempfile
 from tkinter import (
     Tk,
-    StringVar,
-    IntVar,
-    filedialog,
-    messagebox,
-    Menu,
-    TclError,
     PhotoImage,
 )
-from tkinter.ttk import Frame, Label, Entry, Button, Checkbutton, Treeview, Notebook
-from threading import Thread
-from diskcache import Cache
 
 # local imports
-from getmyancestors.classes.classes import (
-    Session,
-    Tree,
-    Indi,
-    Fam,
-    Gedcom,
-    EntryWithMenu,
-    FilesToMerge,
-    Merge,
-    SignIn,
-    StartIndis,
-    Options,
-    Download,
+from getmyancestors.classes.gui import (
     FStoGEDCOM,
 )
-
-from getmyancestors.translation import translations
 
 
 def main():
