+++ getmyancestors/getmyancestors.py
@@ -12,20 +12,13 @@ import argparse
 
 # local imports
 import getmyancestors
-from getmyancestors.translation import translations
-from getmyancestors.classes.classes import (
-    Session,
-    Note,
-    Source,
-    Fact,
-    Memorie,
-    Name,
-    Ordinance,
-    Indi,
-    Fam,
+
+from getmyancestors.classes.tree import (
     Tree,
 )
-from getmyancestors.constants import (
+from getmyancestors.classes.session import Session
+from getmyancestors.classes.translation import translations
+from getmyancestors.classes.constants import (
     FACT_TAGS,
     FACT_EVEN,
     MAX_PERSONS,
