+++ setup.cfg
@@ -13,7 +13,7 @@ classifiers =
   Programming Language :: Python :: 3 :: Only
 
 [options]
-packages = getmyancestors
+packages=getmyancestors, getmyancestors.classes
 install_requires =
   babelfish==0.6.0
   diskcache==5.2.1
